+++ src/routes/package.ts
@@ -5,7 +5,7 @@ import type {Src_Json} from '@ryanatkn/gro/src_json.js';
 
 export const package_json = {
 	name: '@ryanatkn/fuz',
-	version: '0.96.0',
+	version: '0.96.2',
 	description: 'Svelte UI library',
 	icon: '🧶',
 	public: true,
@@ -304,7 +304,7 @@ export const package_json = {
 
 export const src_json = {
 	name: '@ryanatkn/fuz',
-	version: '0.96.0',
+	version: '0.96.2',
 	modules: {
 		'./package.json': {path: 'package.json', declarations: []},
 		'./Alert.svelte': {path: 'Alert.svelte', declarations: []},