+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ryanatkn/fuz
 
+## 0.96.2
+
+### Patch Changes
+
+- hoist `repo_name` out of `Library_Content` ([a7e6c77](https://github.com/ryanatkn/fuz/commit/a7e6c77))
+
+## 0.96.1
+
+### Patch Changes
+
+- allow customizing the library breadcrumb children ([a521537](https://github.com/ryanatkn/fuz/commit/a521537))
+- fix `Package_Summary` homepage url ([a521537](https://github.com/ryanatkn/fuz/commit/a521537))
+
 ## 0.96.0
 
 ### Minor Changes